requirements.txt
streamlit==1.28.0
tensorflow-cpu==2.15.0
opencv-python-headless==4.8.1.78
pillow==10.0.0
matplotlib==3.7.2
numpy==1.24.3
